+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Config to enable additional logging for loading and syncing parts and traits
+- Tier 3 analyzer catalyst tag (no items in it by default)
 ### Fixed
 - Traits not loading on OS X